requirements.txt
streamlit
pandas
plotly
python-docx
numpy
transformers
python-dotenv
langchain
langchain-groq
torch
python-docx
docx